requirements.txt

whitenoise
arabic-reshaper==2.1.1
asgiref==3.2.10
Babel==2.8.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.19
django-admin-material==0.1.0
django-allauth==0.42.0
django-easy-pdf==0.1.1
django-environ==0.4.5
django-formtools==2.2
django-two-factor-auth==1.12.1
django-import-export==2.3.0
django-otp==1.0.1
django-otp-yubikey==1.0.0
django-phonenumber-field==3.0.1
django-simple-history==2.12.0
django-two-factor-auth==1.12.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
inflection==0.3.1
jdcal==1.4.1
json2html==1.3.0
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
pdfkit==0.6.1
pep8==1.7.1
phonenumbers==8.12.11
Pillow==8.1.2
psycopg2-binary
pycparser==2.20
pycryptodome==3.9.8
PyJWT==2.0.1
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.65
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
tinycss2==1.1.0
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
yapf==0.30.0
YubiOTP==1.0.0.post1
